A simple tableview exercise
This exercise was to display basic data onto the screen, connect the
Table View Controller code with the storyboard, and create one section
of an array of cars.

This exercise was just to get my feet wet, so I’m assuming 99.99% of
the time, apps won’t look like this.